+++ src/ambition/version_details.cc
@@ -25,6 +25,7 @@ std::vector<VersionUpdates> versionDetails = {
           _("Added a steal report confirmation screen with a Spy-estimated escape chance.  This estimate varies in accuracy over time and based on the spy's skill."),
           _("Show the surplus (or deficit) number of peasants that you have on the Villages Report.  This is how many extra peasants you currently have over the amount you need to exactly balance food production and consumption."),
           _("Increased bribe choices to include $2500 and $3500."),
+          _("Added more food purchasing quantity and price options."),
         },
       },
       {